Update PrimaryNavigation props usage and increase internal spacing
The PrimaryNavigation component was previously using inline styles
to conditionally render the position of the navigation menu. This
produced code that was difficult to maintain and complicated the usage
of media queries. This update will pass the props to the component
directly so that it may be used within the styled component's definition
instead. Moreover, the internal spacing of the component was increased.